Moving locked file list to LocksView
